requirements.txt
Flask==2.3.2
pipreqs==0.4.13
python-dotenv==1.0.0
requests==2.31.0
